Refactor routes into modular structure with separate route files
